Clean Up Preexisting Build Issues
